"Version" request supported
Server version request is now handled in the chat via the command
“version”. Created a method in Context that retrieves a Version #. The
interface BasicView was amended to have a getVersion() method. Added an
instantiated version # in Model.java to represent the version number
along with a method there to access it. The NetworkCode.java file was
amended to have two new codes: get server info request and get server
info response, for the purposes of handling version requests.