"Uptime" command supported
Server uptime request is now handled in the chat via the command
“Uptime”. Created a method in Context that retrieves server uptime.

The interface BasicView was amended to have a getUptime() method. Added
an instantiated Time variable in Model.java to represent the time the
server has been up along with a method there to access it.

 The NetworkCode.java file was amended to have two new codes: get
server uptime request and get server uptime response, for the purposes
of handling uptime requests.