Added Tokenized Input
Added a class to replace the Scanner class that now handles lines of
input from the command line and tokenizes it if input is contained
within quotation marks. It also throws exceptions at mishandled input
and lets the client know in advance.

The Tokenizer class parses through input and inserts it into a list for
later use by Chat.java. It throws exceptions if the input appears to be
unclear, which Chat.java handles with a message.