Uptime print update
Edited it so that when uptime is called in Chat, it returns how long
the server has been running in days hours minutes and seconds. This
included adding a method in the Time class to format seconds to this
format.

Unrelated, deleted the Scanner class imports since they are not being
used anymore!